test(node): Collapse mysql span streaming tests into same suite
