--- dev-packages/node-integration-tests/suites/tracing/mysql-streamed/instrument.mjs
@@ -1,10 +1,0 @@
-import * as Sentry from '@sentry/node';
-import { loggingTransport } from '@sentry-internal/node-integration-tests';
-
-Sentry.init({
-  dsn: 'https://public@dsn.ingest.sentry.io/1337',
-  release: '1.0',
-  tracesSampleRate: 1.0,
-  transport: loggingTransport,
-  traceLifecycle: 'stream',
-});
